+++ driver/src/common/xaie_helper_internal.h
@@ -48,8 +48,8 @@ u8 _XAie_CountTrailingZeros(u32 value);
 u8 _XAie_IsDeviceGenAIE4(u8 DevGen);
 u8 _XAie_IsDeviceGenSupportDualApp(u8 DevGen);
 u8 _XAie_IsTileResourceInSharedAddrSpace(u8 DevGen, u8 AppMode, u8 TileType);
-u8 _XAie_GetMaxElementValue(u8 DevGen, u8 TileType, u8 AppMode,
-		u8 elementValue);
+u16 _XAie_GetMaxElementValue(u8 DevGen, u8 TileType, u8 AppMode,
+		u16 elementValue);
 u64 _XAie_ChangeRegisterSpace(u8 devGen, u64 regOffset);
 u8 _XAie_IsUcModulePresent(XAie_DevInst* DevInst, u8 TileType);
 u8 _XAie_DmaGetMaxNumChannels(XAie_DevInst *DevInst, const XAie_DmaMod *DmaMod,
